Don't explicitly enable CFI in libavc.
Previously there was no other sanitization, so in order for the
blacklist to be picked up CFI had to be enabled. With integer_overflow
enabled, we no longer need it enabled explicitly in the build file.
Since libavc is a static library, both CFI and non-CFI variants will
be built and linked accordingly.

If libavc is used as a shared library in the future, we'll need to
explicitly re-enable CFI in the build file.

Bug: 63926619
Test: Library builds.
Change-Id: I2646e1422c29073d2e622ded1a9074426b186130