Fix macOS build.
There's no <malloc.h> on Darwin. Normally I'd say "prefer <stdlib.h>",
but in this case it looks like the code actually uses memalign. There
is no memalign on Darwin. You'll need to switch to posix_memalign if
you actually care about building this.

For now, the easiest is to just exclude macOS.

Test: treehugger
Change-Id: Id8a11cd141f5c4bd30ce95b61ac64759e17732fc